utilities/requirements.txt
pyfastx==0.8.4
pymongo==3.12.1
pyranges==0.0.115
python_dateutil==2.8.2
PyVCF==0.6.8
requests==2.26.0
